requirements.txt
python-dotenv==1.0.0
pyspark==3.5.0
psycopg2-binary==2.9.9
sentence-transformers==2.2.2
scikit-learn==1.4.0
pandas==2.1.4
numpy==1.26.3
torch==2.2.0
transformers==4.37.0
matplotlib==3.8.2
seaborn==0.13.0
tqdm==4.66.1
requests==2.31.0
pytest==7.4.3
python-dateutil==2.8.2
